Enable 64-bit Darwin LeakSanitizer by default on AddressSanitizer builds
Summary:
LSan on Darwin was originally disabled by default for a couple reasons:
1) To give the opportunity to test thoroughly
2) The performance wasn't great

In terms of the first item, I've been running LSan on a large internal project
(10s of thousands of files of c++ and objc) for about a month and things are working well,
the full test suite for the project runs and passes with LSan enabled on Darwin.

The performance issues have been largely mitigated, and the slowdown over stock ASan
is similar to Linux (10-15% on the tests I've looked at).

Reviewers: kcc, kubamracek, alekseyshl

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D35443